refactor: checkout process and moved createCheckoutSession logic to separate file
